requirements_test.txt
# Strictly for tests
pytest-homeassistant-custom-component